fix produk rating coloumn;

+++ database/migrations/2017_04_09_053527_create_produks_table.php
@@ -27,7 +27,7 @@ class CreateProduksTable extends Migration
             $table->integer('view')->nullable();
             $table->integer('stok');
             $table->boolean('antar')->nullable();
-            $table->integer('rating')->nullable();
+            $table->double('rating')->nullable();
             $table->integer('aktif');
             $table->timestamps();
         });
